+++ pywin32_postinstall.py
@@ -1,4 +1,4 @@
-from win32.lib.pywin32_postinstall import main
+from win32.lib.postinstall.pywin32_postinstall import main
 
 if __name__ == "__main__":
     main()
